Added a new feature: AUTO THREADS,
The bot will listen to message events emmitted for certain selected channels, and will start threads for each message in that channel.

Two new commands:

toggleAutoThread which toggles the mode of auto thread (on or off) off also deletes all selected channels for auto thread

addAutoThreadChannel as the name suggests, it adds a channel to the list of channels that the bot should listen to.